Asynchronous and deferred JavaScript execution
Deferred execution <script defer>
Simply put: delaying script execution until the HTML parser has
finished. A positive effect of this attribute is that the DOM will be
available for your script. However, since not every browser supports
defer yet, don’t rely on it!

Asynchronous execution <script async>
Don’t care when the script will be available? Asynchronous is the best
of both worlds: HTML parsing may continue and the script will be
executed as soon as it’s ready. I’d recommend this for scripts such as
Google Analytics.